+++ mono/project_manager/admin.py
@@ -173,3 +173,10 @@ class InviteAdmin(admin.ModelAdmin):
         'accepted_by',
         'accepted_at',
     ]
+
+
+@admin.register(models.Space)
+class SpaceAdmin(admin.ModelAdmin):
+    list_display = ['name', 'order', 'created_by', 'created_at', 'active']
+    list_filter = ('name', 'created_by', 'created_at', 'active')
+    search_fields = ['name']